make Discovered dependent on autoConnect.
and make QNAM uses sessions, so we get the desired effect of
our connection dialog popping up when someone tries to make
a connection/get, etc when there is no connection.

This forces the connection dialog to popup for all
but autoconnectable services.

It also has the effect that QNAM will wait for a
connection in some instances.