Eschew use of QStringLiteral in connman bearer plugin
The string data QStringLiteral generates is stored in the read-only
segment of the compiled object file. However, libqconnmanbearer.so
can be dynamically unloaded from memory and when this happens, the
data may still be referenced by QStrings e.g. inside Qt DBus. Operation
with such string will lead to application crash because of invalid
memory access.

For this reason change all instances of QStringLiteral to QLatin1String,
with an exception of temporaries used for string comparison, which can't
be referenced from outside of the library.

This problem was observed in crash-reporter-autouploader.